fix(es): Address PR review comments
- Remove UseTimeRangeFilter from Rotation interface (reader uses
  UseReadWriteAliases config directly for the shard-pruning filter)
- Rename OpType to WriteOpType for clarity
- Split rotation_test.go into per-implementation test files
- Fix "dated" comment to "time-based" (supports hourly too)
- Remove rolloverFrequency default from NewPeriodicRotation (defaults
  belong at config level)
- Handle partial rotation overrides independently in reader/writer
- Add consolidated time-range design documentation in reader.go
